refactor(preview): change file name
