+++ src/event-picker/preview/_chart.tsx
